wgengine/router: add a firewall-based "default route killswitch".
Empirically, this is the only reliable way to force Windows to switch
over cleanly to a higher-priority default route: block the old one
such that the sockets get I/O errors and are forced to redial.

The code is copied and adapted from wireguard-windows's firewall module,
with modifications to adjust to Tailscale: we condition only on process
name (same as our existing firewall settings), and we allow DNS out
through the old default route (because right now it's tricky for us
to figure out the correct DNS servers to permit).

This is a bit of a temporary hack. The technique is sound, but we should
be doing more to enable smooth switchover in both directions. However,
this change gets us from default routes not working correctly, to
default routes working mostly correctly. Followups will make it cleaner.

Signed-off-by: David Anderson <danderson@tailscale.com>